+++ R/cities_n.R
@@ -1,0 +1,9 @@
+cities_n <- function(n = 1L, fields = FALSE) {
+  idx <- sample(dim(cities)[1L], n)
+  out <- if(fields) {
+    cities[idx, c("long", "lat", "name", "country.etc", "pop",  "capital")] 
+  } else {
+    cities[idx, c("long", "lat")] 
+  }
+  out
+}