+++ data-raw/cities.R
@@ -1,0 +1,4 @@
+cities <- maps::world.cities
+
+
+usethis::use_data(cities)